Отредактировала предыдущую лабу для ЛР5, добавила авторизацию и пользователей
Отредактировала предыдущую лабу для ЛР5, добавила авторизацию и пользователей

+++ delete_os.php
@@ -1,8 +1,12 @@
 <?php
-$mysqli = new mysqli("localhost", "a0611486_root", "root", "a0611486_os") or die ("Невозможно
+include("checks.php");
+require_once 'connect.php';
+$mysqli = new mysqli($host, $user, $password, $database) or die ("Невозможно
 подключиться к серверу");
 $id_os = $_GET['id_os'];
-$result = $mysqli->query("DELETE FROM os WHERE id_os='$id_os'");
-header("Location: index.php");
+if ($_SESSION['type'] == 2)
+    $result = $mysqli->query("DELETE FROM os WHERE id_os ='$id_os'");
+
+header("Location: osAdm.php");
 exit;
 ?>
